refactor: add errors

+++ src/useCases/deleteFile/DeleteFileUseCase.ts
@@ -1,5 +1,6 @@
 import { IFilesRepository } from '../../repositories/IFilesRepository'
 import { inject, injectable } from 'tsyringe'
+import { AppError } from '../../errors/AppError'
 
 @injectable()
 export class DeleteFileUseCase {
@@ -9,6 +10,9 @@ export class DeleteFileUseCase {
   ) {}
 
   async execute(key: string): Promise<void> {
+    const file = await this.filesRepository.findByKey(key)
+    if (!file) throw new AppError('The provided key is invalid.')
+
     await this.filesRepository.deleteByKey(key)
   }
 }
